alternative-ots-implementation-eks-iac/pipeline-registry/ops/bin/requirements.txt
boto3==1.9.109
PyYAML==3.13
yamllint==1.15.0
